requirements.txt
streamlit
chromadb
pandas
pdfplumber
langchain-text-splitters
sentence-transformers
torch
pathlib